fix: use file offset for section and annotation location.
Sections are always indexed using an offset in MOBI rawml
stream. We were extracting location of annotations using an
offset after converting the raw stream into string. This was
leading to incorrect merge of sections and annotations in
the writer pipeline.

Fix is to use a single addressing for location in both section
and annotation. This is possible if we parse the html of book
section by section and derive the annotation location relative
to the section.